# Conflicts:
#	.gitattributes
#	.gitignore
#	bibc/renum/fetsco.c
#	bibc/wscript
#	bibcxx/IOManager/MedFilePointer.cxx
#	bibcxx/PythonBindings/ContactPairingInterface.cxx
#	bibcxx/PythonBindings/LogicalUnitManager.h
#	bibcxx/wscript
#	bibfor/include/threading_interfaces.h
#	bibfor/third_party_interf/bresels.F90
#	bibfor/third_party_interf/breselsqp.F90
#	bibfor/third_party_interf/breselu.F90
#	bibfor/third_party_interf/verifels.F90
#	bibfor/third_party_interf/verifelu.F90
#	bibfor/wscript
#	data/profile.bat.tmpl
#	data/run_aster.bat
#	data/run_aster_for_ctest.bat
#	data/run_ctest.bat
#	data/wscript
#	pixi.lock
#	pixi.toml
#	run_aster/run_aster_main.py
#	run_aster/run_ctest_main.py
#	run_aster/utils.py
#	waftools/mathematics.py
#	waftools/med_cfg.py
#	waftools/python_cfg.py
#	waftools/wafutils.py
#	wscript

+++ bibfor/include/asterc/aster_petsc_initialize.h
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2020 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -21,6 +21,6 @@
 interface
     subroutine aster_petsc_initialize(myopt, ierr)
         character(len=2500) :: myopt
-        integer, intent(out) :: ierr
+        integer(kind=8), intent(out) :: ierr
     end subroutine aster_petsc_initialize
 end interface
