+++ bibfor/include/asterc/lihdfv.h
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2017 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -20,8 +20,8 @@
 !
 interface
     subroutine lihdfv(major, minor, patch)
-        integer, intent(out) :: major
-        integer, intent(out) :: minor
-        integer, intent(out) :: patch
+        integer(kind=8), intent(out) :: major
+        integer(kind=8), intent(out) :: minor
+        integer(kind=8), intent(out) :: patch
     end subroutine lihdfv
 end interface
