+++ bibfor/include/asterf_mumps.h
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2023 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -29,7 +29,7 @@
 #   include "dmumps_struc.h"
 #   include "zmumps_struc.h"
 
-    integer :: nmxins
+    integer(kind=8) :: nmxins
     parameter (nmxins=50)
     character(len=1) :: roucs(nmxins), precs(nmxins)
     character(len=4) :: etams(nmxins)
