+++ bibfor/include/asterfort/crsvfm.h
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2021 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -24,11 +24,11 @@
               character(len=*) :: matasz
               character        :: prec
               character        :: rank
-              integer          :: pcpiv
+              integer(kind=8)          :: pcpiv
               character(len=*) :: usersmz
               real(kind=8)     :: blreps
               character(len=*) :: renumz
-              integer          :: redmpi
+              integer(kind=8)          :: redmpi
             end subroutine crsvfm
        
           end interface
