+++ bibfor/include/asterfort/filtering.h
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2022 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -22,9 +22,9 @@ interface
     subroutine filtering(ltypr, k, lmnsy, jvale, jvale2, rfiltre, rmin, rmax, &
                      nfilt1, nfilt2, nfilt3, nzloc, iok)
         aster_logical, intent(in) :: ltypr, lmnsy
-        integer, intent(in) :: k, jvale, jvale2
+        integer(kind=8), intent(in) :: k, jvale, jvale2
         real(kind=8), intent(in) :: rfiltre, rmin, rmax
-        integer, intent(inout) :: nfilt1, nfilt2, nfilt3, nzloc
+        integer(kind=8), intent(inout) :: nfilt1, nfilt2, nfilt3, nzloc
         integer(kind=4), intent(inout) :: iok(*)
     end subroutine filtering
 end interface
