+++ bibfor/include/asterfort/jerazo.h
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2017 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -21,7 +21,7 @@
           interface 
             subroutine jerazo(nomlu,ni,i1)
               character(len=*), intent(in) :: nomlu
-              integer, intent(in) :: ni
-              integer, intent(in) :: i1
+              integer(kind=8), intent(in) :: ni
+              integer(kind=8), intent(in) :: i1
             end subroutine jerazo
           end interface
