+++ bibfor/include/asterfort/jjldyn.h
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2017 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -20,8 +20,8 @@
 !
 interface
     subroutine jjldyn(imode, lmin, ltot)
-        integer :: imode
-        integer :: lmin
-        integer :: ltot
+        integer(kind=8) :: imode
+        integer(kind=8) :: lmin
+        integer(kind=8) :: ltot
     end subroutine jjldyn
 end interface
