+++ bibfor/include/asterfort/rscrsd.h
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2020 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -19,6 +19,6 @@
 interface
     subroutine rscrsd(baseZ, resultNameZ, resultTypeZ, nbStore)
         character(len=*), intent(in) :: baseZ, resultNameZ, resultTypeZ
-        integer, intent(in) :: nbStore
+        integer(kind=8), intent(in) :: nbStore
     end subroutine rscrsd
 end interface
