+++ bibfor/include/asterfort/uttcpr.h
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2017 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -20,7 +20,7 @@
 !
 interface
     subroutine uttcpr(nommes, nbv, temps)
-        integer :: nbv
+        integer(kind=8) :: nbv
         character(len=*) :: nommes
         real(kind=8) :: temps(nbv)
     end subroutine uttcpr
